:recycle: Update chat functions
